+++ src/yafowil/bootstrap/__init__.py
@@ -1,5 +1,6 @@
 import os
 from yafowil.base import factory
+from yafowil.bootstrap.common import configure_factory
 
 
 resourcedir = os.path.join(os.path.dirname(__file__), 'resources')
@@ -20,6 +21,11 @@ css = [{
 
 
 def register():
-    import common
     factory.register_theme('bootstrap', 'yafowil.bootstrap',
                            resourcedir, js=js, css=css)
+
+
+def configure():
+    # only configure factory if not suppressed explicit
+    if not os.environ.get('TESTRUN_MARKER'):
+        configure_factory()